Add Cartfile and remove carthage copy phase
Add explicit dependencies using Cartfile so Carthage can detect Macaw dependency when bootstraping / updating.

Remove copy framework phase of Carthage as it should be added in the final Application using frameworks, not the library itself. This copy phase of Macaw was leading to have twice the framework in final App.